chore: revert TS SDK changes (#46)
# Summary

The @FuelLabs/sdk-ts are looking to migrate our API documentation to an
external host (Vercel). As part of this, we applied the following
changes to the GitHub action to make the possible:

- #43 
- #45 

The above changes have caused our CI to fail. We're exploring a
different route to migrate our API docs, therefore, this PR reverts the
above two PR's.